Wed, Aug 2, 2023, 4:58 PM +03:00

+++ sentimental-readability/readability.py
@@ -8,20 +8,20 @@ total_comma = 0
 total_word = 1  # count by space + 1 ot start with 1
 total_sentence = 0  # count by . or !
 
-    for character in text:
-        if character == ',':
-            total_comma += 1
+for character in text:
+    if character == ',':
+        total_comma += 1
 
-        if character == ',' or character == ' ':
-            total_word += 1
-        elif character == '.' or character == '!' or character == ':' or character == '?':
-            total_sentence += 1
-        elif character != "'":
-            total_letter += 1
+    if character == ',' or character == ' ':
+        total_word += 1
+    elif character == '.' or character == '!' or character == ':' or character == '?':
+        total_sentence += 1
+    elif character != "'":
+        total_letter += 1
 
-    L = (total_letter / (total_word - total_comma)) * 100
-    S = (total_sentence / (total_word - total_comma)) * 100
-    actual_grade = (0.0588 * L) - (0.296 * S) - 15.8
+L = (total_letter / (total_word - total_comma)) * 100
+S = (total_sentence / (total_word - total_comma)) * 100
+actual_grade = (0.0588 * L) - (0.296 * S) - 15.8
 
 # print(total_letter)
 # print(total_word)
@@ -30,7 +30,7 @@ total_sentence = 0  # count by . or !
 # # print(L)
 # print(S)
 # print(actual_grade)
-actual_grade = round(actual_grade)
+actual_grade = int(round(actual_grade))
 
 if actual_grade > 16:
     print("Grade 16+")
